Base Learn

+++ src/ArraysExercise.sol
@@ -3,19 +3,19 @@ pragma solidity ^0.8.23;
 
 contract ArraysExercise {
     // Declare state variables to store arrays of numbers, timestamps, and senders
-    uint[] numbers = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]; // Array of numbers initialized with values
-    uint[] timestamps; // Dynamic array to store timestamps
+    uint256[] numbers = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]; // Array of numbers initialized with values
+    uint256[] timestamps; // Dynamic array to store timestamps
     address[] senders; // Dynamic array to store sender addresses
 
     uint256 constant Y2K = 946702800; // Constant representing the Unix timestamp for the year 2000
 
     // Function to retrieve the array of numbers
-    function getNumbers() external view returns (uint[] memory) {
+    function getNumbers() external view returns (uint256[] memory) {
         // Create a memory array to hold the numbers
-        uint[] memory results = new uint[](numbers.length);
+        uint256[] memory results = new uint256[](numbers.length);
 
         // Copy the numbers from the state array to the memory array
-        for(uint i=0; i<numbers.length; i++) {
+        for (uint256 i = 0; i < numbers.length; i++) {
             results[i] = numbers[i];
         }
 
@@ -29,16 +29,16 @@ contract ArraysExercise {
     }
 
     // Function to append new numbers to the numbers array
-    function appendToNumbers(uint[] calldata _toAppend) public {
+    function appendToNumbers(uint256[] calldata _toAppend) public {
         // Iterate through the array to be appended
-        for (uint i = 0; i < _toAppend.length; i++) {
+        for (uint256 i = 0; i < _toAppend.length; i++) {
             // Push each element of the array to be appended to the numbers array
             numbers.push(_toAppend[i]);
         }
     }
 
     // Function to save a timestamp along with the sender's address
-    function saveTimestamp(uint _unixTimestamp) public {
+    function saveTimestamp(uint256 _unixTimestamp) public {
         // Push the timestamp and sender's address to their respective arrays
         timestamps.push(_unixTimestamp);
         senders.push(msg.sender);
@@ -50,7 +50,7 @@ contract ArraysExercise {
         uint256 counter = 0;
 
         // Count the number of timestamps after Y2K
-        for (uint i = 0; i < timestamps.length; i++) {
+        for (uint256 i = 0; i < timestamps.length; i++) {
             if (timestamps[i] > Y2K) {
                 counter++;
             }
@@ -64,7 +64,7 @@ contract ArraysExercise {
         uint256 index = 0;
 
         // Iterate through timestamps and senders arrays to extract elements after Y2K
-        for (uint i = 0; i < timestamps.length; i++) {
+        for (uint256 i = 0; i < timestamps.length; i++) {
             if (timestamps[i] > Y2K) {
                 timestampsAfterY2K[index] = timestamps[i];
                 sendersAfterY2K[index] = senders[i];
